kbupd-config: add hostname DNS lookup checking
The 'validate' subcommand has a new option, that is disabled by
default:

    -d, --check-dns-hostnames    Check that the hostnames resolve with DNS

When activated, kbupd-config will attempt a DNS lookup on all
"hostPort" fields within a configuration file.

For frontends, the hostPort field appears within the
FrontendPartitionConfig section.

For replicas, the hostPort field appears in the ReplicaEnclaveConfig
section and optionally in the ReplicaSourcePartitionConfig section.

This commit also includes positive and negative test cases.